a major refactor
