+++ REQUIRE
@@ -1,2 +1,2 @@
-julia 0.5
+julia 0.6
 Compat 0.19
