--- src/precompile.jl
@@ -1,251 +1,0 @@
-function _precompile_()
-    ccall(:jl_generating_output, Cint, ()) == 1 || return nothing
-    precompile(BandedMatrices.gbmm!, (Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float32, BandedMatrices.BandedMatrix{Float32}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float64}, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float32}, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_axpy!, (Base.Complex{Float64}, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.banded_axpy!, (Float64, BandedMatrices.BandedMatrix{Int64}, Base.SubArray{Int64, 2, BandedMatrices.BandedMatrix{Int64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.gbmm!, (Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Float32, BandedMatrices.BandedMatrix{Float32},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Float64, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.banded_axpy!, (Float64, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_axpy!, (Float64, BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.Complex{Float64}, BandedMatrices.BandedMatrix{Base.Complex{Float64}},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.Complex{Float32}, BandedMatrices.BandedMatrix{Base.Complex{Float32}},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Int64, 2, BandedMatrices.BandedMatrix{Int64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Float64}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.gbmm!, (Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Float32}, Float32, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.Complex{Float64}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float64}, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.Complex{Float64}, BandedMatrices.BandedMatrix{Base.Complex{Float64}},))
-    precompile(BandedMatrices.gbmm!, (Float32, BandedMatrices.BandedMatrix{Float32}, BandedMatrices.BandedMatrix{Float32}, Float32, BandedMatrices.BandedMatrix{Float32},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_axpy!, (Float64, BandedMatrices.BandedMatrix{Float64}, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, BandedMatrices.BandedMatrix{Float64}, BandedMatrices.BandedMatrix{Float64}, Float64, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float32}, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Base.Complex{Float32}, BandedMatrices.BandedMatrix{Base.Complex{Float32}},))
-    precompile(BandedMatrices.banded_dense_axpy!, (Float64, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Array{Int64, 1}, Array{Int64, 1}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_dense_axpy!, (Float64, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Array{Int64, 1}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_dense_axpy!, (Float64, BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Array{Int64, 1}, Array{Int64, 1}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Float32}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, BandedMatrices.BandedMatrix{Float32}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_dense_axpy!, (Float64, BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Array{Int64, 1}, Base.Colon}, false},))
-    precompile(BandedMatrices.banded_axpy!, (Float64, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_axpy!, (Float64, BandedMatrices.BandedMatrix{Float32}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.banded_dense_axpy!, (Float64, BandedMatrices.BandedMatrix{Float32}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Array{Int64, 1}, Array{Int64, 1}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Float32},))
-    precompile(BandedMatrices.banded_A_mul_B!, (Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Base.Complex{Float64}},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Base.Complex{Float32}},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Array{Float32, 2}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.banded_axpy!, (Base.Complex{Float64}, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, BandedMatrices.BandedMatrix{Base.Complex{Float64}},))
-    precompile(BandedMatrices.banded_axpy!, (Float32, BandedMatrices.BandedMatrix{Float32}, BandedMatrices.BandedMatrix{Float32},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Float32}, BandedMatrices.BandedMatrix{Float32}, BandedMatrices.BandedMatrix{Float32},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 1, Array{Base.Complex{Float64}, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 1, Array{Base.Complex{Float64}, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_dense_axpy!, (Float64, BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, Array{Float64, 2}, Tuple{Array{Int64, 1}, Array{Int64, 1}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, BandedMatrices.BandedMatrix{Base.Complex{Float64}},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.bandinds, (BandedMatrices.BandedMatrix{Float32}, Int64,))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Float64}, BandedMatrices.BandedMatrix{Float64}, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Base.Complex{Float32}}, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, BandedMatrices.BandedMatrix{Base.Complex{Float32}},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.checkbounds, (BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Int64, Int64,))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.bandwidths, (BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.banded_A_mul_B!, (Array{Base.Complex{Float64}, 1}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 1, Array{Base.Complex{Float64}, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Array{Base.Complex{Float64}, 2}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.bandinds, (BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Int64,))
-    precompile(BandedMatrices.bandinds, (BandedMatrices.BandedMatrix{Float64}, Int64,))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 1, Array{Base.Complex{Float64}, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 1, Array{Base.Complex{Float64}, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.banded_A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 1, Array{Base.Complex{Float64}, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Array{Base.Complex{Float64}, 1},))
-    precompile(BandedMatrices.bandinds, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Int64,))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Array{Base.Complex{Float32}, 2}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Type{Base.Complex{Float32}}, Int64, Int64, Int64, Int64,))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Array{Float64, 2}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Type{Base.Complex{Float64}}, Int64, Int64, Int64, Int64,))
-    precompile(BandedMatrices.normalize!, (Int64, Ptr{Float64},))
-    precompile(BandedMatrices.gbmm!, (Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.Complex{Float32}, Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Type{Float64}, Int64, Int64, Int64, Int64,))
-    precompile(BandedMatrices.banded_A_mul_B!, (Array{Float64, 1}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.banded_A_mul_B!, (Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Array{Float64, 1},))
-    precompile(BandedMatrices.bandwidth, (Array{Base.Complex{Float32}, 2}, Int64,))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.bandwidth, (Array{Float32, 2}, Int64,))
-    precompile(BandedMatrices.gbmv!, (Char, Int64, Int64, Int64, Float32, Ptr{Float32}, Int64, Int64, Ptr{Float32}, Int64, Float32, Ptr{Float32}, Int64,))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix}, Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmv!, (Char, Int64, Int64, Int64, Base.Complex{Float32}, Ptr{Base.Complex{Float32}}, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Base.Complex{Float32}, Ptr{Base.Complex{Float32}}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Array{Float64, 2}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Array{Base.Complex{Float64}, 2}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Array{Int64, 2}, Int64,))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.bandwidths, (BandedMatrices.BandedMatrix{Base.Complex{Float64}},))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Type{Float32}, Int64, Int64, Int64, Int64,))
-    precompile(BandedMatrices.banded_A_mul_B!, (Array{Base.Complex{Float64}, 1}, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Array{Base.Complex{Float64}, 1},))
-    precompile(BandedMatrices.Anon_Bnon_C_gbmv!, (Float64, Float64, Int64, Int64, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.gbmv!, (Char, Int64, Int64, Int64, Base.Complex{Float64}, Ptr{Base.Complex{Float64}}, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Base.Complex{Float64}, Ptr{Base.Complex{Float64}}, Int64,))
-    precompile(BandedMatrices.bandwidths, (BandedMatrices.BandedMatrix{Float32},))
-    precompile(BandedMatrices.banded_A_mul_B!, (Array{Float64, 1}, BandedMatrices.BandedMatrix{Float64}, Array{Float64, 1},))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix{Base.Complex{Float64}}}, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.Atop_Bmid_Ctop_gbmv!, (Float32, Float32, Int64, Int64, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.normalize!, (Int64, Ptr{Base.Complex{Float64}},))
-    precompile(BandedMatrices.Anon_Bnon_C_gbmv!, (Base.Complex{Float64}, Base.Complex{Float64}, Int64, Int64, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.gbmv!, (Char, Int64, Int64, Int64, Float64, Ptr{Float64}, Int64, Int64, Ptr{Float64}, Int64, Float64, Ptr{Float64}, Int64,))
-    precompile(BandedMatrices.checkbounds, (BandedMatrices.BandedMatrix{Int64}, Int64, Int64,))
-    precompile(BandedMatrices.Amid_Bmid_Cmid_gbmv!, (Float32, Float32, Int64, Int64, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.Anon_Bnon_C_gbmv!, (Base.Complex{Float32}, Base.Complex{Float32}, Int64, Int64, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Type{Int64}, Int64, Int64, Int64, Int64,))
-    precompile(BandedMatrices.checkbounds, (BandedMatrices.BandedMatrix{Float64}, Int64, Int64,))
-    precompile(BandedMatrices.Anon_Bnon_C_gbmv!, (Float32, Float32, Int64, Int64, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.checkbounds, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Int64, Int64,))
-    precompile(BandedMatrices.checksquare, (BandedMatrices.BandedMatrix{Float32},))
-    precompile(BandedMatrices.checkbounds, (BandedMatrices.BandedMatrix{Float32}, Int64, Int64,))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Base.Complex{Float64}, 1, Array{Base.Complex{Float64}, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Array{Base.Complex{Float64}, 1},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Array{Float64, 1},))
-    precompile(BandedMatrices.getindex, (BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Int64, Int64,))
-    precompile(BandedMatrices.checksquare, (BandedMatrices.BandedMatrix{Base.Complex{Float64}},))
-    precompile(BandedMatrices.inbands_setindex!, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Base.Complex{Float64}, Int64, Int64,))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix{Base.Complex{Float32}}}, BandedMatrices.BandedMatrix{Base.Complex{Float32}},))
-    precompile(BandedMatrices.Amid_Bmid_Cmid_gbmv!, (Base.Complex{Float32}, Base.Complex{Float32}, Int64, Int64, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.Atop_Bmid_Ctop_gbmv!, (Base.Complex{Float32}, Base.Complex{Float32}, Int64, Int64, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.Atop_Bmid_Ctop_gbmv!, (Float64, Float64, Int64, Int64, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix{Float32}}, BandedMatrices.BandedMatrix{Float32},))
-    precompile(BandedMatrices.A_mul_B!, (Array{Base.Complex{Float64}, 1}, Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Base.Complex{Float64}, 1, Array{Base.Complex{Float64}, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.A_mul_B!, (Array{Float64, 1}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.getindex, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Int64, Int64,))
-    precompile(BandedMatrices.checksquare, (BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.Amid_Bmid_Cmid_gbmv!, (Float64, Float64, Int64, Int64, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix{Int64}}, BandedMatrices.BandedMatrix{Int64},))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix{Float64}}, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.getindex, (BandedMatrices.BandedMatrix{Int64}, Int64, Int64,))
-    precompile(BandedMatrices.A11_Btop_Ctop_gbmv!, (Base.Complex{Float64}, Base.Complex{Float64}, Int64, Int64, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.getindex, (BandedMatrices.BandedMatrix{Float64}, Int64, Int64,))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix{Base.Complex{Float64}}}, BandedMatrices.BandedMatrix{Base.Complex{Float64}},))
-    precompile(BandedMatrices.inbands_setindex!, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Float64, Int64, Int64,))
-    precompile(BandedMatrices.Atop_Bmid_Ctop_gbmv!, (Base.Complex{Float64}, Base.Complex{Float64}, Int64, Int64, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.A11_Btop_Ctop_gbmv!, (Float64, Float64, Int64, Int64, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.Amid_Bmid_Cmid_gbmv!, (Base.Complex{Float64}, Base.Complex{Float64}, Int64, Int64, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64, Ptr{Base.Complex{Float64}}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.inbands_setindex!, (BandedMatrices.BandedMatrix{Float32}, Float32, Int64, Int64,))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.getindex, (BandedMatrices.BandedMatrix{Float32}, Int64, Int64,))
-    precompile(BandedMatrices.bzeros, (Type{Base.Complex{Float64}}, Int64, Int64, Int64, Int64,))
-    precompile(BandedMatrices.inbands_setindex!, (BandedMatrices.BandedMatrix{Float32}, Float64, Int64, Int64,))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.checkdimensions, (Tuple{Int64}, Tuple{Int64},))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Array{Int64, 2}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.rowrange, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Int64,))
-    precompile(BandedMatrices.rowrange, (BandedMatrices.BandedMatrix{Float64}, Int64,))
-    precompile(BandedMatrices.bandwidths, (BandedMatrices.BandedMatrix{Base.Complex{Float32}},))
-    precompile(BandedMatrices.size, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Int64,))
-    precompile(BandedMatrices.bzeros, (Type{Float32}, Int64, Int64, Int64, Int64,))
-    precompile(BandedMatrices.bzeros, (Type{Float64}, Int64, Int64, Int64, Int64,))
-    precompile(BandedMatrices.size, (BandedMatrices.BandedMatrix{Float64}, Int64,))
-    precompile(BandedMatrices.size, (BandedMatrices.BandedMatrix{Float32}, Int64,))
-    precompile(BandedMatrices.full, (BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.A11_Btop_Ctop_gbmv!, (Base.Complex{Float32}, Base.Complex{Float32}, Int64, Int64, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64, Ptr{Base.Complex{Float32}}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.A11_Btop_Ctop_gbmv!, (Float32, Float32, Int64, Int64, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64, Ptr{Float32}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.checkband, (BandedMatrices.BandedMatrix{Float32}, Int64,))
-    precompile(BandedMatrices.checkband, (BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Int64,))
-    precompile(BandedMatrices.checkband, (BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Int64,))
-    precompile(BandedMatrices.checkband, (BandedMatrices.BandedMatrix{Float64}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Base.Complex{Float64}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float64}}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Base.Complex{Float32}, 2, BandedMatrices.BandedMatrix{Base.Complex{Float32}}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Float32, 2, BandedMatrices.BandedMatrix{Float32}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Int64,))
-
-    # from ApproxFun
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A11_Btop_Ctop_gbmv!, (Float64, Float64, Int64, Int64, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.banded_axpy!, (Float64, BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Float64, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, BandedMatrices.BandedMatrix{Float64}, BandedMatrices.BandedMatrix{Float64}, Float64, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Float64}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.A_mul_B!, (BandedMatrices.BandedMatrix{Float64}, BandedMatrices.BandedMatrix{Float64}, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, BandedMatrices.BandedMatrix{Float64}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.A_mul_B!, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.banded_A_mul_B!, (Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Array{Float64, 1},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false},))
-    precompile(BandedMatrices.banded_A_mul_B!, (Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Array{Float64, 2}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.bandinds, (BandedMatrices.BandedMatrix{Float64}, Int64,))
-    precompile(BandedMatrices.normalize!, (Int64, Ptr{Float64},))
-    precompile(BandedMatrices.bandwidths, (BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.bandwidth, (Array{Float64, 2}, Int64,))
-    precompile(BandedMatrices.banded_A_mul_B!, (Array{Float64, 1}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 1, Array{Float64, 1}, Tuple{Base.UnitRange{Int64}}, true},))
-    precompile(BandedMatrices.convert, (Type{AbstractArray{Float64, 2}}, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.banded_A_mul_B!, (Array{Float64, 1}, BandedMatrices.BandedMatrix{Float64}, Array{Float64, 1},))
-    precompile(BandedMatrices.Amid_Bmid_Cmid_gbmv!, (Float64, Float64, Int64, Int64, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.Anon_Bnon_C_gbmv!, (Float64, Float64, Int64, Int64, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.gbmv!, (Char, Int64, Int64, Int64, Float64, Ptr{Float64}, Int64, Int64, Ptr{Float64}, Int64, Float64, Ptr{Float64}, Int64,))
-    precompile(BandedMatrices.Atop_Bmid_Ctop_gbmv!, (Float64, Float64, Int64, Int64, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64, Ptr{Float64}, Int64, Int64, Int64,))
-    precompile(BandedMatrices.convert, (Type{BandedMatrices.BandedMatrix{Float64}}, BandedMatrices.BandedMatrix{Float64},))
-    precompile(BandedMatrices.bandshift, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.Type, (Type{BandedMatrices.BandedMatrix}, Type{Float64}, Int64, Int64, Int64, Int64,))
-    precompile(BandedMatrices.checkbounds, (BandedMatrices.BandedMatrix{Float64}, Int64, Int64,))
-    precompile(BandedMatrices.rowrange, (BandedMatrices.BandedMatrix{Float64}, Int64,))
-    precompile(BandedMatrices.gbmm!, (Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Float64, Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false},))
-    precompile(BandedMatrices.checkband, (BandedMatrices.BandedMatrix{Float64}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.Colon, Base.UnitRange{Int64}}, false}, Int64,))
-    precompile(BandedMatrices.bandwidth, (Base.SubArray{Float64, 2, BandedMatrices.BandedMatrix{Float64}, Tuple{Base.UnitRange{Int64}, Base.Colon}, false}, Int64,))
-end
